nlp in the field eye tracking
Based on existing eye-tracking algorithms, images can be generated that display the trajectory of a person's gaze while viewing certain information. The background used is a slide from the monitor screen during the examination, over which fixations and saccades are displayed.
The task is to convert visual eye-tracking data (fixations, saccades, changes in pupil diameter) into a structured textual description in natural language.